requirements.txt
Django
djangorestframework
djangorestframework-jwt==1.10.0
djoser==0.6.0
psycopg2

